experiments/veo3-item-consistency/pyproject.toml
[project]
name = "veo-facial-consistency"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.25.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
]
